feature/accounts #8
1. 최종 테스트 및 QA

(1) 일부 코드 수정

-   Class 상속으로 인해 api 결과에 "result" 키가 포함되지행
(2) 전체 기능에 대한 테스트 완료

-   완료 일시: 2024-05-06 15:07

+++ sparta_news/accounts/views.py
@@ -2,7 +2,7 @@ from rest_framework.status import *
 from rest_framework.views import APIView
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.exceptions import AuthenticationFailed
 from rest_framework_simplejwt.views import TokenObtainPairView
 from rest_framework_simplejwt.serializers import TokenBlacklistSerializer
 from rest_framework_simplejwt.exceptions import TokenError
@@ -20,18 +20,30 @@ ACCOUNTS_MNG = ACCOUNTS_MODEL.objects
 @api_view(["POST"])
 def signup(request):
 
+    result = {
+        "result": False,
+        "msg": None
+    }
+    status_code = HTTP_400_BAD_REQUEST
+
     # POST 입력 데이터 Serialization
     serializer = AccountSerializers(data=request.data,
                                     many=False)
 
     # POST 입력값 검증
-    serializer.is_valid(raise_exception=True)
+    if serializer.is_valid():
+        # 새 가입 계정 생성
+        data = serializer.save()
+        result.pop("msg")
+        result["user"] = data
+        status_code = HTTP_200_OK
 
-    # 새 가입 계정 생성
-    data = serializer.save()
+    else:
+        result["msg"] = serializer.errors
 
     # API 처리 결과 반환
-    return Response(data={"result": True, "user": data})
+    return Response(data=result,
+                    status=status_code)
 
 
 # 회원 로그인
@@ -39,23 +51,35 @@ class AccountSignInView(TokenObtainPairView):
 
     def post(self, request, *args, **kwargs):
 
-        # 데이터 추출 / 계정 검증 및 token 반환
-        data = super().post(request, *args, **kwargs)
-
-        # refresh 토큰 추출
-        r_token = data.data.get("refresh")
-
-        # 로그인 계정 last_date 및 r_token 업데이트
-        update_last_login(username=request.data.get("username"),
-                          r_token = r_token)
-
-        # 발급된 access_token 반환
-        data = {
-            "result": True,
-            "access_token": data.data.get("access")
+        result = {
+            "result": False,
+            "msg": None
         }
-        return Response(data=data,
-                        status=HTTP_200_OK)
+        status_code = HTTP_401_UNAUTHORIZED
+
+        # 데이터 추출 / 계정 검증 및 token 반환
+        try:
+            data = super().post(request, *args, **kwargs)
+
+        except AuthenticationFailed as e:
+            result["msg"] = "계정 정보가 일치하지 않습니다."
+
+        else:
+            # refresh 토큰 추출
+            r_token = data.data.get("refresh")
+
+            # 로그인 계정 last_date 및 r_token 업데이트
+            update_last_login(username=request.data.get("username"),
+                              r_token = r_token)
+
+            # 발급된 access_token 반환
+            result["result"] = True
+            result["access_token"] = data.data.get("access")
+            result.pop("msg")
+            status_code = HTTP_200_OK
+
+        return Response(data=result,
+                        status=status_code)
 
 
 # 회원 로그아웃
@@ -90,29 +114,37 @@ def signout(request):
 
 class AccountsDetailView(APIView):
 
-    permission_classes = [IsAuthenticated]
+    def check_token(self, request) -> bool:
+        if request.auth is None:
+            return False
+
+        return True
 
     # 회원 정보 조회
     def get(self, request, account_id: int):
 
         result = {
             "result": False,
-            "msg": "계정이 존재하지 않습니다."
+            "msg": None
         }
-        status_code = HTTP_400_BAD_REQUEST
+        status_code = HTTP_401_UNAUTHORIZED
 
-        try:
-            user = ACCOUNTS_MNG.get(id=account_id)
-
-        except ACCOUNTS_MODEL.DoesNotExist:
-            pass
+        if not self.check_token(request=request):
+            result["msg"] = "로그인 사용자의 유효한 토큰이 입력되지 않았습니다."
 
         else:
-            serializer = AccountSerializers(instance=user, many=False)
-            result["result"] = True
-            result["user"] = serializer.get_data()
-            result.pop("msg")
-            status_code = HTTP_200_OK
+            try:
+                user = ACCOUNTS_MNG.get(id=account_id)
+
+            except ACCOUNTS_MODEL.DoesNotExist:
+                result["msg"] = "사용자가 존재하지 않습니다."
+
+            else:
+                serializer = AccountSerializers(instance=user, many=False)
+                result["result"] = True
+                result["user"] = serializer.get_data()
+                result.pop("msg")
+                status_code = HTTP_200_OK
 
         return Response(data=result,
                         status=status_code)
@@ -122,25 +154,35 @@ class AccountsDetailView(APIView):
 
         result = {
             "result": False,
-            "msg": "다른 계정의 정보 변경을 시도했습니다."
+            "msg": None
         }
         status_code = HTTP_401_UNAUTHORIZED
 
-        if request.user.id == account_id:
+        if not self.check_token(request=request):
+            result["msg"] = "로그인 사용자의 유효한 토큰이 입력되지 않았습니다."
 
-            serializer = AccountsModifySerializers(data=request.data,
-                                                   instance=ACCOUNTS_MNG.get(id=account_id),
-                                                   many=False,
-                                                   partial=True)
+        else:
+            if request.user.id == account_id:
+                serializer = AccountsModifySerializers(data=request.data,
+                                                       instance=ACCOUNTS_MNG.get(id=account_id),
+                                                       many=False,
+                                                       partial=True)
 
-            serializer.is_valid(raise_exception=True)
-            serializer.save()
+                if serializer.is_valid():
+                    serializer.save()
 
-            result["result"] = True
-            result["user"] = AccountSerializers(instance=ACCOUNTS_MNG.get(id=account_id),
-                                                many=False).get_data()
-            result.pop("msg")
-            status_code = HTTP_200_OK
+                    result["result"] = True
+                    result["user"] = AccountSerializers(instance=ACCOUNTS_MNG.get(id=account_id),
+                                                        many=False).get_data()
+                    result.pop("msg")
+                    status_code = HTTP_200_OK
+
+                else:
+                    result["msg"] = serializer.errors
+                    status_code = HTTP_400_BAD_REQUEST
+
+            else:
+                result["msg"] = "다른 계정의 변경을 시도했습니다."
 
         return Response(data=result,
                         status=status_code)
@@ -150,28 +192,37 @@ class AccountsDetailView(APIView):
 
         result = {
             "result": False,
-            "msg": "다른 계정의 비밀번호 변경을 시도했습니다."
+            "msg": None
         }
         status_code = HTTP_401_UNAUTHORIZED
 
-        if request.user.id == account_id:
+        if not self.check_token(request=request):
+            result["msg"] = "로그인 사용자의 유효한 토큰이 입력되지 않았습니다."
 
-            serializer = AccountsPasswordChangeSerializer(data=request.data,
-                                                          instance=ACCOUNTS_MNG.get(id=account_id),
-                                                          many=False)
+        else:
+            if request.user.id == account_id:
+                serializer = AccountsPasswordChangeSerializer(data=request.data,
+                                                              instance=ACCOUNTS_MNG.get(id=account_id),
+                                                              many=False)
 
-            serializer.is_valid(raise_exception=True)
-            user = just_authenticate(request=request,
-                                     username=request.user.username,
-                                     password=request.data.get("pre_password"))
+                if serializer.is_valid():
+                    user = just_authenticate(request=request,
+                                             username=request.user.username,
+                                             password=request.data.get("pre_password"))
 
-            if user is not None and serializer.save(request=request):
-                result["result"] = True
-                result.pop("msg")
-                status_code = HTTP_204_NO_CONTENT
+                    if user is not None and serializer.save(request=request):
+                        result["result"] = True
+                        result.pop("msg")
+                        status_code = HTTP_204_NO_CONTENT
+
+                    else:
+                        result["msg"] = "비밀번호가 일치하지 않습니다."
+
+                else:
+                    result["msg"] = serializer.errors
 
             else:
-                result["msg"] = "비밀번호가 일치하지 않습니다."
+                result["msg"] = "다른 계정의 비밀번호 변경을 시도했습니다."
 
         return Response(data=result,
                         status=status_code)
@@ -181,27 +232,37 @@ class AccountsDetailView(APIView):
 
         result = {
             "result": False,
-            "msg": "다른 계정의 탈퇴를 시도하였습니다."
+            "msg": None
         }
         status_code = HTTP_401_UNAUTHORIZED
 
-        if request.user.id == account_id:
+        if not self.check_token(request=request):
+            result["msg"] = "로그인 사용자의 유효한 토큰이 입력되지 않았습니다."
 
-            user = ACCOUNTS_MNG.get(id=account_id)
-            input_pw = request.data.get("password")
+        else:
+            if request.user.id == account_id:
 
-            if just_authenticate(request=request,
-                                 username=user.username,
-                                 password=input_pw):
+                user = ACCOUNTS_MNG.get(id=account_id)
+                input_pw = request.data.get("password")
 
-                user.delete()
-                result["result"] = True
-                result.pop("msg")
-                status_code = HTTP_204_NO_CONTENT
+                if input_pw is not None:
+                    if just_authenticate(request=request,
+                                         username=user.username,
+                                         password=input_pw):
+
+                        user.delete()
+                        result["result"] = True
+                        result.pop("msg")
+                        status_code = HTTP_204_NO_CONTENT
+
+                    else:
+                        result["msg"] = "계정 정보가 일치하지 않습니다."
+
+                else:
+                    result["msg"] = "탈퇴할 계정의 비밀번호를 입력해야합니다. (key: password)"
 
             else:
-                result["msg"] = "계정 정보가 일치하지 않습니다."
+                result["msg"] = "다른 계정의 탈퇴를 시도했습니다."
 
         return Response(data=result,
                         status=status_code)
-
